refactor(api): deprecate POST /schooling-registration-user-associations/possibilities
in favor of POST /sco-organization-learners/possibilities
