+++ api/lib/application/organization-learner-user-associations/index.js
@@ -40,38 +40,6 @@ exports.register = async function (server) {
       },
     },
     {
-      method: 'PUT',
-      path: '/api/schooling-registration-user-associations/possibilities',
-      config: {
-        auth: false,
-        handler: organizationLearnerUserAssociationController.generateUsername,
-        validate: {
-          options: {
-            allowUnknown: true,
-          },
-          payload: Joi.object({
-            data: {
-              attributes: {
-                'first-name': Joi.string().empty(Joi.string().regex(/^\s*$/)).required(),
-                'last-name': Joi.string().empty(Joi.string().regex(/^\s*$/)).required(),
-                birthdate: Joi.date().format('YYYY-MM-DD').raw().required(),
-                'campaign-code': Joi.string().empty(Joi.string().regex(/^\s*$/)).required(),
-              },
-            },
-          }),
-          failAction: (request, h) => {
-            return sendJsonApiError(new UnprocessableEntityError('Un des champs saisis n’est pas valide.'), h);
-          },
-        },
-        notes: [
-          '- Elle permet de savoir si un élève identifié par son nom, prénom et date de naissance est inscrit à ' +
-            "l'organisation détenant la campagne. Cet élève n'est, de plus, pas encore associé à l'organisation.",
-        ],
-        tags: ['api', 'organizationLearnerUserAssociation'],
-      },
-    },
-
-    {
       method: 'PATCH',
       path: '/api/organizations/{id}/schooling-registration-user-associations/{schoolingRegistrationId}',
       config: {
