+++ api/lib/application/sco-organization-learners/index.js
@@ -125,6 +125,69 @@ exports.register = async function (server) {
         tags: ['api', 'sco-organization-learners'],
       },
     },
+    {
+      method: 'PUT',
+      path: '/api/schooling-registration-user-associations/possibilities',
+      config: {
+        auth: false,
+        handler: scoOrganizationLearnerController.generateUsername,
+        validate: {
+          options: {
+            allowUnknown: true,
+          },
+          payload: Joi.object({
+            data: {
+              attributes: {
+                'first-name': Joi.string().empty(Joi.string().regex(/^\s*$/)).required(),
+                'last-name': Joi.string().empty(Joi.string().regex(/^\s*$/)).required(),
+                birthdate: Joi.date().format('YYYY-MM-DD').raw().required(),
+                'campaign-code': Joi.string().empty(Joi.string().regex(/^\s*$/)).required(),
+              },
+            },
+          }),
+          failAction: (request, h) => {
+            return sendJsonApiError(new UnprocessableEntityError('Un des champs saisis n’est pas valide.'), h);
+          },
+        },
+        notes: [
+          '- Elle permet de savoir si un élève identifié par son nom, prénom et date de naissance est inscrit à ' +
+            "l'organisation détenant la campagne. Cet élève n'est, de plus, pas encore associé à l'organisation.",
+          "- L'usage de cette route est **dépréciée** en faveur de /api/sco-organization-learners/possibilities",
+        ],
+        tags: ['api', 'organizationLearnerUserAssociation'],
+      },
+    },
+    {
+      method: 'PUT',
+      path: '/api/sco-organization-learners/possibilities',
+      config: {
+        auth: false,
+        handler: scoOrganizationLearnerController.generateUsername,
+        validate: {
+          options: {
+            allowUnknown: true,
+          },
+          payload: Joi.object({
+            data: {
+              attributes: {
+                'first-name': Joi.string().empty(Joi.string().regex(/^\s*$/)).required(),
+                'last-name': Joi.string().empty(Joi.string().regex(/^\s*$/)).required(),
+                birthdate: Joi.date().format('YYYY-MM-DD').raw().required(),
+                'campaign-code': Joi.string().empty(Joi.string().regex(/^\s*$/)).required(),
+              },
+            },
+          }),
+          failAction: (request, h) => {
+            return sendJsonApiError(new UnprocessableEntityError('Un des champs saisis n’est pas valide.'), h);
+          },
+        },
+        notes: [
+          '- Elle permet de savoir si un élève identifié par son nom, prénom et date de naissance est inscrit à ' +
+            "l'organisation détenant la campagne. Cet élève n'est, de plus, pas encore associé à l'organisation.",
+        ],
+        tags: ['api', 'sco-organization-learners'],
+      },
+    },
   ]);
 };
