+++ api/lib/infrastructure/serializers/jsonapi/sco-organization-learner-serializer.js
@@ -6,4 +6,10 @@ module.exports = {
       attributes: ['lastName', 'firstName', 'birthdate'],
     }).serialize(scoOrganizationLearner);
   },
+
+  serializeWithUsernameGeneration(scoOrganizationLearner) {
+    return new Serializer('sco-organization-learner', {
+      attributes: ['lastName', 'firstName', 'birthdate', 'username'],
+    }).serialize(scoOrganizationLearner);
+  },
 };
