+++ api/tests/acceptance/application/sco-organization-learner-controller_test.js
@@ -1166,4 +1166,793 @@ describe('Acceptance | Controller | sco-organization-learners', function () {
       });
     });
   });
+
+  describe('PUT /api/schooling-registration-user-associations/possibilities', function () {
+    let options;
+    let user;
+    let organization;
+    let organizationLearner;
+    let campaignCode;
+
+    beforeEach(async function () {
+      organization = databaseBuilder.factory.buildOrganization({
+        isManagingStudents: true,
+        type: 'SCO',
+      });
+      campaignCode = databaseBuilder.factory.buildCampaign({
+        organizationId: organization.id,
+      }).code;
+      user = databaseBuilder.factory.buildUser();
+      organizationLearner = databaseBuilder.factory.buildOrganizationLearner({
+        organizationId: organization.id,
+        firstName: user.firstName,
+        lastName: user.lastName,
+        userId: null,
+        nationalStudentId: 'nsi123ABC',
+      });
+      await databaseBuilder.commit();
+
+      options = {
+        method: 'PUT',
+        url: '/api/schooling-registration-user-associations/possibilities',
+        headers: {},
+        payload: {
+          data: {
+            attributes: {
+              'campaign-code': campaignCode,
+              'first-name': organizationLearner.firstName,
+              'last-name': organizationLearner.lastName,
+              birthdate: organizationLearner.birthdate,
+            },
+          },
+        },
+      };
+    });
+
+    describe('Success case', function () {
+      it('should return the organizationLearner linked to the user and a 200 status code response', async function () {
+        // given
+        const expectedResult = {
+          data: {
+            attributes: {
+              birthdate: '2005-08-05',
+              'campaign-code': 'ABC456TTY',
+              'first-name': 'Billy',
+              'last-name': 'TheKid',
+              username: 'billy.thekid0508',
+            },
+            type: 'schooling-registration-user-associations',
+          },
+        };
+
+        // when
+        const response = await server.inject(options);
+
+        // then
+        expect(response.statusCode).to.equal(200);
+        expect(response.result).to.deep.equal(expectedResult);
+      });
+    });
+
+    describe('Error cases', function () {
+      context('when no organizationLearner can be associated because birthdate does not match', function () {
+        it('should respond with a 404 - Not Found', async function () {
+          // given
+          options.payload.data.attributes.birthdate = '1990-03-01';
+
+          // when
+          const response = await server.inject(options);
+
+          // then
+          expect(response.statusCode).to.equal(404);
+          expect(response.result.errors[0].detail).to.equal(
+            'There were no organizationLearners matching with organization and birthdate'
+          );
+        });
+      });
+
+      context('when no organizationLearner found to associate because names does not match', function () {
+        it('should respond with a 404 - Not Found', async function () {
+          // given
+          options.payload.data.attributes['first-name'] = 'wrong firstName';
+
+          // when
+          const response = await server.inject(options);
+
+          // then
+          expect(response.statusCode).to.equal(404);
+          expect(response.result.errors[0].detail).to.equal('There were no organizationLearners matching with names');
+        });
+      });
+
+      context('when organizationLearner is already associated in the same organization', function () {
+        it('should return an organization learner already linked error (short code S51 when account with email)', async function () {
+          // given
+          const userWithEmailOnly = databaseBuilder.factory.buildUser({
+            username: null,
+            email: 'john.harry@example.net',
+          });
+          const organizationLearner = databaseBuilder.factory.buildOrganizationLearner({
+            organizationId: organization.id,
+            userId: userWithEmailOnly.id,
+          });
+          await databaseBuilder.commit();
+
+          const expectedResponse = {
+            status: '409',
+            code: 'ACCOUNT_WITH_EMAIL_ALREADY_EXIST_FOR_THE_SAME_ORGANIZATION',
+            title: 'Conflict',
+            detail: 'Un compte existe déjà pour l‘élève dans le même établissement.',
+            meta: { shortCode: 'S51', value: 'j***@example.net' },
+          };
+
+          options.headers.authorization = generateValidRequestAuthorizationHeader(userWithEmailOnly.id);
+          options.payload.data = {
+            attributes: {
+              'campaign-code': campaignCode,
+              'first-name': organizationLearner.firstName,
+              'last-name': organizationLearner.lastName,
+              birthdate: organizationLearner.birthdate,
+            },
+          };
+
+          // when
+          const response = await server.inject(options);
+
+          // then
+          expect(response.statusCode).to.equal(409);
+          expect(response.result.errors[0]).to.deep.equal(expectedResponse);
+        });
+
+        it('should return an organization learner already linked error (short code S52 when connected with username)', async function () {
+          // given
+          const userWithUsernameOnly = databaseBuilder.factory.buildUser({
+            username: 'john.harry0702',
+            email: null,
+          });
+          const organizationLearner = databaseBuilder.factory.buildOrganizationLearner({
+            organizationId: organization.id,
+            userId: userWithUsernameOnly.id,
+          });
+          await databaseBuilder.commit();
+
+          const expectedResponse = {
+            status: '409',
+            code: 'ACCOUNT_WITH_USERNAME_ALREADY_EXIST_FOR_THE_SAME_ORGANIZATION',
+            title: 'Conflict',
+            detail: 'Un compte existe déjà pour l‘élève dans le même établissement.',
+            meta: { shortCode: 'S52', value: 'j***.h***2' },
+          };
+
+          options.headers.authorization = generateValidRequestAuthorizationHeader(userWithUsernameOnly.id);
+          options.payload.data = {
+            attributes: {
+              'campaign-code': campaignCode,
+              'first-name': organizationLearner.firstName,
+              'last-name': organizationLearner.lastName,
+              birthdate: organizationLearner.birthdate,
+            },
+          };
+
+          // when
+          const response = await server.inject(options);
+
+          // then
+          expect(response.statusCode).to.equal(409);
+          expect(response.result.errors[0]).to.deep.equal(expectedResponse);
+        });
+
+        it('should return an organization learner already linked error (short code S53 when account with samlId)', async function () {
+          // given
+          const userWithEmailOnly = databaseBuilder.factory.buildUser({
+            username: null,
+            email: null,
+          });
+          databaseBuilder.factory.buildAuthenticationMethod.withGarAsIdentityProvider({
+            externalIdentifier: '12345678',
+            userId: userWithEmailOnly.id,
+          });
+
+          const organizationLearner = databaseBuilder.factory.buildOrganizationLearner({
+            firstName: 'Sam',
+            lastName: 'Lebrave',
+            birthdate: '2015-10-10',
+            nationalStudentId: 'samlebrave',
+            organizationId: organization.id,
+            userId: null,
+          });
+          organizationLearner.userId = userWithEmailOnly.id;
+          await databaseBuilder.commit();
+
+          const expectedResponse = {
+            status: '409',
+            code: 'ACCOUNT_WITH_GAR_ALREADY_EXIST_FOR_THE_SAME_ORGANIZATION',
+            title: 'Conflict',
+            detail: 'Un compte existe déjà pour l‘élève dans le même établissement.',
+            meta: { shortCode: 'S53', value: null },
+          };
+
+          options.headers.authorization = generateValidRequestAuthorizationHeader(userWithEmailOnly.id);
+          options.payload.data = {
+            attributes: {
+              'campaign-code': campaignCode,
+              'first-name': organizationLearner.firstName,
+              'last-name': organizationLearner.lastName,
+              birthdate: organizationLearner.birthdate,
+            },
+          };
+
+          // when
+          const response = await server.inject(options);
+
+          // then
+          expect(response.statusCode).to.equal(409);
+          expect(response.result.errors[0]).to.deep.equal(expectedResponse);
+        });
+      });
+
+      context('when organizationLearner is already associated in others organizations', function () {
+        it('should respond with a 409 - Conflict', async function () {
+          // given
+          const organizationLearnerAlreadyMatched = databaseBuilder.factory.buildOrganizationLearner({
+            birthdate: '2005-05-15',
+            firstName: user.firstName,
+            lastName: user.lastName,
+            organizationId: organization.id,
+            userId: user.id,
+            nationalStudentId: 'coucoucloclo',
+          });
+          await databaseBuilder.commit();
+
+          options.payload.data.attributes['first-name'] = organizationLearnerAlreadyMatched.firstName;
+          options.payload.data.attributes['last-name'] = organizationLearnerAlreadyMatched.lastName;
+          options.payload.data.attributes.birthdate = organizationLearnerAlreadyMatched.birthdate;
+
+          // when
+          const response = await server.inject(options);
+
+          // then
+          expect(response.statusCode).to.equal(409);
+          expect(response.result.errors[0].detail).to.equal(
+            'Un compte existe déjà pour l‘élève dans le même établissement.'
+          );
+        });
+
+        it('should return an organization learner already linked error (short code S61 when account with email)', async function () {
+          // given
+          const userWithEmailOnly = databaseBuilder.factory.buildUser({
+            username: null,
+            email: 'john.harry@example.net',
+          });
+          databaseBuilder.factory.buildOrganizationLearner({
+            nationalStudentId: organizationLearner.nationalStudentId,
+            birthdate: organizationLearner.birthdate,
+            firstName: organizationLearner.firstName,
+            lastName: organizationLearner.lastName,
+            userId: userWithEmailOnly.id,
+          });
+          await databaseBuilder.commit();
+
+          const expectedResponse = {
+            status: '409',
+            code: 'ACCOUNT_WITH_EMAIL_ALREADY_EXIST_FOR_ANOTHER_ORGANIZATION',
+            title: 'Conflict',
+            detail: 'Un compte existe déjà pour l‘élève dans un autre établissement.',
+            meta: { shortCode: 'S61', value: 'j***@example.net' },
+          };
+
+          options.headers.authorization = generateValidRequestAuthorizationHeader(userWithEmailOnly.id);
+          options.payload.data = {
+            attributes: {
+              'campaign-code': campaignCode,
+              'first-name': organizationLearner.firstName,
+              'last-name': organizationLearner.lastName,
+              birthdate: organizationLearner.birthdate,
+            },
+          };
+
+          // when
+          const response = await server.inject(options);
+
+          // then
+          expect(response.statusCode).to.equal(409);
+          expect(response.result.errors[0]).to.deep.equal(expectedResponse);
+        });
+
+        it('should return an organization learner already linked error (short code S62 when connected with username)', async function () {
+          // given
+          const userWithUsernameOnly = databaseBuilder.factory.buildUser({
+            username: 'john.harry0702',
+            email: null,
+          });
+          databaseBuilder.factory.buildOrganizationLearner({
+            nationalStudentId: organizationLearner.nationalStudentId,
+            birthdate: organizationLearner.birthdate,
+            firstName: organizationLearner.firstName,
+            lastName: organizationLearner.lastName,
+            userId: userWithUsernameOnly.id,
+          });
+          await databaseBuilder.commit();
+
+          const expectedResponse = {
+            status: '409',
+            code: 'ACCOUNT_WITH_USERNAME_ALREADY_EXIST_FOR_ANOTHER_ORGANIZATION',
+            title: 'Conflict',
+            detail: 'Un compte existe déjà pour l‘élève dans un autre établissement.',
+            meta: { shortCode: 'S62', value: 'j***.h***2' },
+          };
+
+          options.headers.authorization = generateValidRequestAuthorizationHeader(userWithUsernameOnly.id);
+          options.payload.data = {
+            attributes: {
+              'campaign-code': campaignCode,
+              'first-name': organizationLearner.firstName,
+              'last-name': organizationLearner.lastName,
+              birthdate: organizationLearner.birthdate,
+            },
+          };
+
+          // when
+          const response = await server.inject(options);
+
+          // then
+          expect(response.statusCode).to.equal(409);
+          expect(response.result.errors[0]).to.deep.equal(expectedResponse);
+        });
+
+        it('should return an organization learner already linked error (short code S63 when account with samlId)', async function () {
+          // given
+          const userWithSamlIdOnly = databaseBuilder.factory.buildUser({
+            email: null,
+            username: null,
+          });
+          databaseBuilder.factory.buildAuthenticationMethod.withGarAsIdentityProvider({
+            externalIdentifier: '12345678',
+            userId: userWithSamlIdOnly.id,
+          });
+
+          databaseBuilder.factory.buildOrganizationLearner({
+            nationalStudentId: organizationLearner.nationalStudentId,
+            birthdate: organizationLearner.birthdate,
+            firstName: organizationLearner.firstName,
+            lastName: organizationLearner.lastName,
+            userId: userWithSamlIdOnly.id,
+          });
+          await databaseBuilder.commit();
+
+          const expectedResponse = {
+            status: '409',
+            code: 'ACCOUNT_WITH_GAR_ALREADY_EXIST_FOR_ANOTHER_ORGANIZATION',
+            title: 'Conflict',
+            detail: 'Un compte existe déjà pour l‘élève dans un autre établissement.',
+            meta: { shortCode: 'S63', value: null },
+          };
+
+          options.headers.authorization = generateValidRequestAuthorizationHeader(userWithSamlIdOnly.id);
+          options.payload.data = {
+            attributes: {
+              'campaign-code': campaignCode,
+              'first-name': organizationLearner.firstName,
+              'last-name': organizationLearner.lastName,
+              birthdate: organizationLearner.birthdate,
+            },
+          };
+
+          // when
+          const response = await server.inject(options);
+
+          // then
+          expect(response.statusCode).to.equal(409);
+          expect(response.result.errors[0]).to.deep.equal(expectedResponse);
+        });
+      });
+
+      context('when a field is not valid', function () {
+        it('should respond with a 422 - Unprocessable Entity', async function () {
+          // given
+          options.payload.data.attributes['last-name'] = ' ';
+
+          // when
+          const response = await server.inject(options);
+
+          // then
+          expect(response.statusCode).to.equal(422);
+        });
+      });
+    });
+  });
+
+  describe('PUT /api/sco-organization-learners/possibilities', function () {
+    let options;
+    let user;
+    let organization;
+    let organizationLearner;
+    let campaignCode;
+
+    beforeEach(async function () {
+      organization = databaseBuilder.factory.buildOrganization({
+        isManagingStudents: true,
+        type: 'SCO',
+      });
+      campaignCode = databaseBuilder.factory.buildCampaign({
+        organizationId: organization.id,
+      }).code;
+      user = databaseBuilder.factory.buildUser();
+      organizationLearner = databaseBuilder.factory.buildOrganizationLearner({
+        organizationId: organization.id,
+        firstName: user.firstName,
+        lastName: user.lastName,
+        userId: null,
+        nationalStudentId: 'nsi123ABC',
+      });
+      await databaseBuilder.commit();
+
+      options = {
+        method: 'PUT',
+        url: '/api/sco-organization-learners/possibilities',
+        headers: {},
+        payload: {
+          data: {
+            attributes: {
+              'campaign-code': campaignCode,
+              'first-name': organizationLearner.firstName,
+              'last-name': organizationLearner.lastName,
+              birthdate: organizationLearner.birthdate,
+            },
+          },
+        },
+      };
+    });
+
+    describe('Success case', function () {
+      it('should return the organizationLearner linked to the user and a 200 status code response', async function () {
+        //given
+        const expectedResult = {
+          data: {
+            attributes: {
+              birthdate: '2005-08-05',
+              'first-name': 'Billy',
+              'last-name': 'TheKid',
+              username: 'billy.thekid0508',
+            },
+            type: 'sco-organization-learners',
+          },
+        };
+
+        // when
+        const response = await server.inject(options);
+
+        // then
+        expect(response.statusCode).to.equal(200);
+        expect(response.result).to.deep.equal(expectedResult);
+      });
+    });
+
+    describe('Error cases', function () {
+      context('when no organizationLearner can be associated because birthdate does not match', function () {
+        it('should respond with a 404 - Not Found', async function () {
+          // given
+          options.payload.data.attributes.birthdate = '1990-03-01';
+
+          // when
+          const response = await server.inject(options);
+
+          // then
+          expect(response.statusCode).to.equal(404);
+          expect(response.result.errors[0].detail).to.equal(
+            'There were no organizationLearners matching with organization and birthdate'
+          );
+        });
+      });
+
+      context('when no organizationLearner found to associate because names does not match', function () {
+        it('should respond with a 404 - Not Found', async function () {
+          // given
+          options.payload.data.attributes['first-name'] = 'wrong firstName';
+
+          // when
+          const response = await server.inject(options);
+
+          // then
+          expect(response.statusCode).to.equal(404);
+          expect(response.result.errors[0].detail).to.equal('There were no organizationLearners matching with names');
+        });
+      });
+
+      context('when organizationLearner is already associated in the same organization', function () {
+        it('should return an organization learner already linked error (short code S51 when account with email)', async function () {
+          // given
+          const userWithEmailOnly = databaseBuilder.factory.buildUser({
+            username: null,
+            email: 'john.harry@example.net',
+          });
+          const organizationLearner = databaseBuilder.factory.buildOrganizationLearner({
+            organizationId: organization.id,
+            userId: userWithEmailOnly.id,
+          });
+          await databaseBuilder.commit();
+
+          const expectedResponse = {
+            status: '409',
+            code: 'ACCOUNT_WITH_EMAIL_ALREADY_EXIST_FOR_THE_SAME_ORGANIZATION',
+            title: 'Conflict',
+            detail: 'Un compte existe déjà pour l‘élève dans le même établissement.',
+            meta: { shortCode: 'S51', value: 'j***@example.net' },
+          };
+
+          options.headers.authorization = generateValidRequestAuthorizationHeader(userWithEmailOnly.id);
+          options.payload.data = {
+            attributes: {
+              'campaign-code': campaignCode,
+              'first-name': organizationLearner.firstName,
+              'last-name': organizationLearner.lastName,
+              birthdate: organizationLearner.birthdate,
+            },
+          };
+
+          // when
+          const response = await server.inject(options);
+
+          // then
+          expect(response.statusCode).to.equal(409);
+          expect(response.result.errors[0]).to.deep.equal(expectedResponse);
+        });
+
+        it('should return an organization learner already linked error (short code S52 when connected with username)', async function () {
+          // given
+          const userWithUsernameOnly = databaseBuilder.factory.buildUser({
+            username: 'john.harry0702',
+            email: null,
+          });
+          const organizationLearner = databaseBuilder.factory.buildOrganizationLearner({
+            organizationId: organization.id,
+            userId: userWithUsernameOnly.id,
+          });
+          await databaseBuilder.commit();
+
+          const expectedResponse = {
+            status: '409',
+            code: 'ACCOUNT_WITH_USERNAME_ALREADY_EXIST_FOR_THE_SAME_ORGANIZATION',
+            title: 'Conflict',
+            detail: 'Un compte existe déjà pour l‘élève dans le même établissement.',
+            meta: { shortCode: 'S52', value: 'j***.h***2' },
+          };
+
+          options.headers.authorization = generateValidRequestAuthorizationHeader(userWithUsernameOnly.id);
+          options.payload.data = {
+            attributes: {
+              'campaign-code': campaignCode,
+              'first-name': organizationLearner.firstName,
+              'last-name': organizationLearner.lastName,
+              birthdate: organizationLearner.birthdate,
+            },
+          };
+
+          // when
+          const response = await server.inject(options);
+
+          // then
+          expect(response.statusCode).to.equal(409);
+          expect(response.result.errors[0]).to.deep.equal(expectedResponse);
+        });
+
+        it('should return an organization learner already linked error (short code S53 when account with samlId)', async function () {
+          // given
+          const userWithEmailOnly = databaseBuilder.factory.buildUser({
+            username: null,
+            email: null,
+          });
+          databaseBuilder.factory.buildAuthenticationMethod.withGarAsIdentityProvider({
+            externalIdentifier: '12345678',
+            userId: userWithEmailOnly.id,
+          });
+
+          const organizationLearner = databaseBuilder.factory.buildOrganizationLearner({
+            firstName: 'Sam',
+            lastName: 'Lebrave',
+            birthdate: '2015-10-10',
+            nationalStudentId: 'samlebrave',
+            organizationId: organization.id,
+            userId: null,
+          });
+          organizationLearner.userId = userWithEmailOnly.id;
+          await databaseBuilder.commit();
+
+          const expectedResponse = {
+            status: '409',
+            code: 'ACCOUNT_WITH_GAR_ALREADY_EXIST_FOR_THE_SAME_ORGANIZATION',
+            title: 'Conflict',
+            detail: 'Un compte existe déjà pour l‘élève dans le même établissement.',
+            meta: { shortCode: 'S53', value: null },
+          };
+
+          options.headers.authorization = generateValidRequestAuthorizationHeader(userWithEmailOnly.id);
+          options.payload.data = {
+            attributes: {
+              'campaign-code': campaignCode,
+              'first-name': organizationLearner.firstName,
+              'last-name': organizationLearner.lastName,
+              birthdate: organizationLearner.birthdate,
+            },
+          };
+
+          // when
+          const response = await server.inject(options);
+
+          // then
+          expect(response.statusCode).to.equal(409);
+          expect(response.result.errors[0]).to.deep.equal(expectedResponse);
+        });
+      });
+
+      context('when organizationLearner is already associated in others organizations', function () {
+        it('should respond with a 409 - Conflict', async function () {
+          // given
+          const organizationLearnerAlreadyMatched = databaseBuilder.factory.buildOrganizationLearner({
+            birthdate: '2005-05-15',
+            firstName: user.firstName,
+            lastName: user.lastName,
+            organizationId: organization.id,
+            userId: user.id,
+            nationalStudentId: 'coucoucloclo',
+          });
+          await databaseBuilder.commit();
+
+          options.payload.data.attributes['first-name'] = organizationLearnerAlreadyMatched.firstName;
+          options.payload.data.attributes['last-name'] = organizationLearnerAlreadyMatched.lastName;
+          options.payload.data.attributes.birthdate = organizationLearnerAlreadyMatched.birthdate;
+
+          // when
+          const response = await server.inject(options);
+
+          // then
+          expect(response.statusCode).to.equal(409);
+          expect(response.result.errors[0].detail).to.equal(
+            'Un compte existe déjà pour l‘élève dans le même établissement.'
+          );
+        });
+
+        it('should return an organization learner already linked error (short code S61 when account with email)', async function () {
+          // given
+          const userWithEmailOnly = databaseBuilder.factory.buildUser({
+            username: null,
+            email: 'john.harry@example.net',
+          });
+          databaseBuilder.factory.buildOrganizationLearner({
+            nationalStudentId: organizationLearner.nationalStudentId,
+            birthdate: organizationLearner.birthdate,
+            firstName: organizationLearner.firstName,
+            lastName: organizationLearner.lastName,
+            userId: userWithEmailOnly.id,
+          });
+          await databaseBuilder.commit();
+
+          const expectedResponse = {
+            status: '409',
+            code: 'ACCOUNT_WITH_EMAIL_ALREADY_EXIST_FOR_ANOTHER_ORGANIZATION',
+            title: 'Conflict',
+            detail: 'Un compte existe déjà pour l‘élève dans un autre établissement.',
+            meta: { shortCode: 'S61', value: 'j***@example.net' },
+          };
+
+          options.headers.authorization = generateValidRequestAuthorizationHeader(userWithEmailOnly.id);
+          options.payload.data = {
+            attributes: {
+              'campaign-code': campaignCode,
+              'first-name': organizationLearner.firstName,
+              'last-name': organizationLearner.lastName,
+              birthdate: organizationLearner.birthdate,
+            },
+          };
+
+          // when
+          const response = await server.inject(options);
+
+          // then
+          expect(response.statusCode).to.equal(409);
+          expect(response.result.errors[0]).to.deep.equal(expectedResponse);
+        });
+
+        it('should return an organization learner already linked error (short code S62 when connected with username)', async function () {
+          // given
+          const userWithUsernameOnly = databaseBuilder.factory.buildUser({
+            username: 'john.harry0702',
+            email: null,
+          });
+          databaseBuilder.factory.buildOrganizationLearner({
+            nationalStudentId: organizationLearner.nationalStudentId,
+            birthdate: organizationLearner.birthdate,
+            firstName: organizationLearner.firstName,
+            lastName: organizationLearner.lastName,
+            userId: userWithUsernameOnly.id,
+          });
+          await databaseBuilder.commit();
+
+          const expectedResponse = {
+            status: '409',
+            code: 'ACCOUNT_WITH_USERNAME_ALREADY_EXIST_FOR_ANOTHER_ORGANIZATION',
+            title: 'Conflict',
+            detail: 'Un compte existe déjà pour l‘élève dans un autre établissement.',
+            meta: { shortCode: 'S62', value: 'j***.h***2' },
+          };
+
+          options.headers.authorization = generateValidRequestAuthorizationHeader(userWithUsernameOnly.id);
+          options.payload.data = {
+            attributes: {
+              'campaign-code': campaignCode,
+              'first-name': organizationLearner.firstName,
+              'last-name': organizationLearner.lastName,
+              birthdate: organizationLearner.birthdate,
+            },
+          };
+
+          // when
+          const response = await server.inject(options);
+
+          // then
+          expect(response.statusCode).to.equal(409);
+          expect(response.result.errors[0]).to.deep.equal(expectedResponse);
+        });
+
+        it('should return an organization learner already linked error (short code S63 when account with samlId)', async function () {
+          // given
+          const userWithSamlIdOnly = databaseBuilder.factory.buildUser({
+            email: null,
+            username: null,
+          });
+          databaseBuilder.factory.buildAuthenticationMethod.withGarAsIdentityProvider({
+            externalIdentifier: '12345678',
+            userId: userWithSamlIdOnly.id,
+          });
+
+          databaseBuilder.factory.buildOrganizationLearner({
+            nationalStudentId: organizationLearner.nationalStudentId,
+            birthdate: organizationLearner.birthdate,
+            firstName: organizationLearner.firstName,
+            lastName: organizationLearner.lastName,
+            userId: userWithSamlIdOnly.id,
+          });
+          await databaseBuilder.commit();
+
+          const expectedResponse = {
+            status: '409',
+            code: 'ACCOUNT_WITH_GAR_ALREADY_EXIST_FOR_ANOTHER_ORGANIZATION',
+            title: 'Conflict',
+            detail: 'Un compte existe déjà pour l‘élève dans un autre établissement.',
+            meta: { shortCode: 'S63', value: null },
+          };
+
+          options.headers.authorization = generateValidRequestAuthorizationHeader(userWithSamlIdOnly.id);
+          options.payload.data = {
+            attributes: {
+              'campaign-code': campaignCode,
+              'first-name': organizationLearner.firstName,
+              'last-name': organizationLearner.lastName,
+              birthdate: organizationLearner.birthdate,
+            },
+          };
+
+          // when
+          const response = await server.inject(options);
+
+          // then
+          expect(response.statusCode).to.equal(409);
+          expect(response.result.errors[0]).to.deep.equal(expectedResponse);
+        });
+      });
+
+      context('when a field is not valid', function () {
+        it('should respond with a 422 - Unprocessable Entity', async function () {
+          // given
+          options.payload.data.attributes['last-name'] = ' ';
+
+          // when
+          const response = await server.inject(options);
+
+          // then
+          expect(response.statusCode).to.equal(422);
+        });
+      });
+    });
+  });
 });
