+++ api/tests/integration/application/sco-organization-learners/index_test.js
@@ -13,6 +13,9 @@ describe('Integration | Application | Route | sco-organization-learners', functi
     sinon
       .stub(scoOrganizationLearnerController, 'reconcileScoOrganizationLearnerAutomatically')
       .callsFake((request, h) => h.response('ok').code(204));
+    sinon
+      .stub(scoOrganizationLearnerController, 'generateUsername')
+      .callsFake((request, h) => h.response('ok').code(200));
 
     httpTestServer = new HttpTestServer();
     await httpTestServer.register(moduleUnderTest);
@@ -466,4 +469,367 @@ describe('Integration | Application | Route | sco-organization-learners', functi
     });
   });
 
+  describe('PUT /api/schooling-registration-user-associations/possibilities', function () {
+    const method = 'PUT';
+    const url = '/api/schooling-registration-user-associations/possibilities';
+
+    it('should exist', async function () {
+      // given
+      const payload = {
+        data: {
+          attributes: {
+            'first-name': 'Robert',
+            'last-name': 'Smith',
+            birthdate: '2012-12-12',
+            'campaign-code': 'RESTRICTD',
+          },
+        },
+      };
+
+      // when
+      const response = await httpTestServer.request(method, url, payload);
+
+      // then
+      expect(response.statusCode).to.equal(200);
+    });
+
+    it('should succeed when there is a space', async function () {
+      // given
+      const payload = {
+        data: {
+          attributes: {
+            'first-name': 'Robert ',
+            'last-name': 'Smith',
+            birthdate: '2012-12-12',
+            'campaign-code': 'RESTRICTD',
+          },
+        },
+      };
+
+      // when
+      const response = await httpTestServer.request(method, url, payload);
+
+      // then
+      expect(response.statusCode).to.equal(200);
+      expect(response.request.payload.data.attributes['first-name']).to.equal('Robert ');
+    });
+
+    it('should return an error when there is no payload', async function () {
+      // when
+      const response = await httpTestServer.request(method, url);
+
+      // then
+      expect(response.statusCode).to.equal(422);
+    });
+
+    it('should return an error when there is an invalid first name attribute in the payload', async function () {
+      // given
+      const INVALID_FIRSTNAME = ' ';
+      const payload = {
+        data: {
+          attributes: {
+            'first-name': INVALID_FIRSTNAME,
+            'last-name': 'Smith',
+            birthdate: '2012-12-12',
+            'campaign-code': 'RESTRICTD',
+          },
+        },
+      };
+
+      // when
+      const response = await httpTestServer.request(method, url, payload);
+
+      // then
+      expect(response.statusCode).to.equal(422);
+    });
+
+    it('should return an error when there is an invalid last name attribute in the payload', async function () {
+      // given
+      const INVALID_LASTNAME = '';
+      const payload = {
+        data: {
+          attributes: {
+            'first-name': 'Robert',
+            'last-name': INVALID_LASTNAME,
+            birthdate: '2012-12-12',
+            'campaign-code': 'RESTRICTD',
+          },
+        },
+      };
+
+      // when
+      const response = await httpTestServer.request(method, url, payload);
+
+      // then
+      expect(response.statusCode).to.equal(422);
+    });
+
+    it('should return an error when there is an invalid a birthdate attribute (with space) in the payload', async function () {
+      // given
+      const INVALID_BIRTHDATE = '2012- 12-12';
+      const payload = {
+        data: {
+          attributes: {
+            'first-name': 'Robert',
+            'last-name': 'Smith',
+            birthdate: INVALID_BIRTHDATE,
+            'campaign-code': 'RESTRICTD',
+          },
+        },
+      };
+
+      // when
+      const response = await httpTestServer.request(method, url, payload);
+
+      // then
+      expect(response.statusCode).to.equal(422);
+    });
+
+    it('should return an error when there is an invalid birthdate attribute (with extra zeros) in the payload', async function () {
+      // given
+      const INVALID_BIRTHDATE = '2012-012-12';
+
+      // when
+      const payload = {
+        data: {
+          attributes: {
+            'first-name': 'Robert',
+            'last-name': 'Smith',
+            birthdate: INVALID_BIRTHDATE,
+            'campaign-code': 'RESTRICTD',
+          },
+        },
+      };
+
+      // when
+      const response = await httpTestServer.request(method, url, payload);
+
+      // then
+      expect(response.statusCode).to.equal(422);
+    });
+
+    it('should return an error when there is an invalid birthdate attribute (not a proper date) in the payload', async function () {
+      // given
+      const INVALID_BIRTHDATE = '1999-99-99';
+      const payload = {
+        data: {
+          attributes: {
+            'first-name': 'Robert',
+            'last-name': 'Smith',
+            birthdate: INVALID_BIRTHDATE,
+            'campaign-code': 'RESTRICTD',
+          },
+        },
+      };
+
+      // when
+      const response = await httpTestServer.request(method, url, payload);
+
+      // then
+      expect(response.statusCode).to.equal(422);
+    });
+
+    it('should return an error when there is an invalid campaign code attribute in the payload', async function () {
+      // given
+      const INVALID_CAMPAIGNCODE = '';
+      const payload = {
+        data: {
+          attributes: {
+            'first-name': 'Robert',
+            'last-name': 'Smith',
+            birthdate: '2012-12-12',
+            'campaign-code': INVALID_CAMPAIGNCODE,
+          },
+        },
+      };
+
+      // when
+      const response = await httpTestServer.request(method, url, payload);
+
+      // then
+      expect(response.statusCode).to.equal(422);
+    });
+  });
+
+  describe('PUT /api/sco-organization-learners/possibilities', function () {
+    const method = 'PUT';
+    const url = '/api/sco-organization-learners/possibilities';
+
+    it('should exist', async function () {
+      // given
+      const payload = {
+        data: {
+          attributes: {
+            'first-name': 'Robert',
+            'last-name': 'Smith',
+            birthdate: '2012-12-12',
+            'campaign-code': 'RESTRICTD',
+          },
+        },
+      };
+
+      // when
+      const response = await httpTestServer.request(method, url, payload);
+
+      // then
+      expect(response.statusCode).to.equal(200);
+    });
+
+    it('should succeed when there is a space', async function () {
+      // given
+      const payload = {
+        data: {
+          attributes: {
+            'first-name': 'Robert ',
+            'last-name': 'Smith',
+            birthdate: '2012-12-12',
+            'campaign-code': 'RESTRICTD',
+          },
+        },
+      };
+
+      // when
+      const response = await httpTestServer.request(method, url, payload);
+
+      // then
+      expect(response.statusCode).to.equal(200);
+      expect(response.request.payload.data.attributes['first-name']).to.equal('Robert ');
+    });
+
+    it('should return an error when there is no payload', async function () {
+      // when
+      const response = await httpTestServer.request(method, url);
+
+      // then
+      expect(response.statusCode).to.equal(422);
+    });
+
+    it('should return an error when there is an invalid first name attribute in the payload', async function () {
+      // given
+      const INVALID_FIRSTNAME = ' ';
+      const payload = {
+        data: {
+          attributes: {
+            'first-name': INVALID_FIRSTNAME,
+            'last-name': 'Smith',
+            birthdate: '2012-12-12',
+            'campaign-code': 'RESTRICTD',
+          },
+        },
+      };
+
+      // when
+      const response = await httpTestServer.request(method, url, payload);
+
+      // then
+      expect(response.statusCode).to.equal(422);
+    });
+
+    it('should return an error when there is an invalid last name attribute in the payload', async function () {
+      // given
+      const INVALID_LASTNAME = '';
+      const payload = {
+        data: {
+          attributes: {
+            'first-name': 'Robert',
+            'last-name': INVALID_LASTNAME,
+            birthdate: '2012-12-12',
+            'campaign-code': 'RESTRICTD',
+          },
+        },
+      };
+
+      // when
+      const response = await httpTestServer.request(method, url, payload);
+
+      // then
+      expect(response.statusCode).to.equal(422);
+    });
+
+    it('should return an error when there is an invalid a birthdate attribute (with space) in the payload', async function () {
+      // given
+      const INVALID_BIRTHDATE = '2012- 12-12';
+      const payload = {
+        data: {
+          attributes: {
+            'first-name': 'Robert',
+            'last-name': 'Smith',
+            birthdate: INVALID_BIRTHDATE,
+            'campaign-code': 'RESTRICTD',
+          },
+        },
+      };
+
+      // when
+      const response = await httpTestServer.request(method, url, payload);
+
+      // then
+      expect(response.statusCode).to.equal(422);
+    });
+
+    it('should return an error when there is an invalid birthdate attribute (with extra zeros) in the payload', async function () {
+      // given
+      const INVALID_BIRTHDATE = '2012-012-12';
+
+      // when
+      const payload = {
+        data: {
+          attributes: {
+            'first-name': 'Robert',
+            'last-name': 'Smith',
+            birthdate: INVALID_BIRTHDATE,
+            'campaign-code': 'RESTRICTD',
+          },
+        },
+      };
+
+      // when
+      const response = await httpTestServer.request(method, url, payload);
+
+      // then
+      expect(response.statusCode).to.equal(422);
+    });
+
+    it('should return an error when there is an invalid birthdate attribute (not a proper date) in the payload', async function () {
+      // given
+      const INVALID_BIRTHDATE = '1999-99-99';
+      const payload = {
+        data: {
+          attributes: {
+            'first-name': 'Robert',
+            'last-name': 'Smith',
+            birthdate: INVALID_BIRTHDATE,
+            'campaign-code': 'RESTRICTD',
+          },
+        },
+      };
+
+      // when
+      const response = await httpTestServer.request(method, url, payload);
+
+      // then
+      expect(response.statusCode).to.equal(422);
+    });
+
+    it('should return an error when there is an invalid campaign code attribute in the payload', async function () {
+      // given
+      const INVALID_CAMPAIGNCODE = '';
+      const payload = {
+        data: {
+          attributes: {
+            'first-name': 'Robert',
+            'last-name': 'Smith',
+            birthdate: '2012-12-12',
+            'campaign-code': INVALID_CAMPAIGNCODE,
+          },
+        },
+      };
+
+      // when
+      const response = await httpTestServer.request(method, url, payload);
+
+      // then
+      expect(response.statusCode).to.equal(422);
+    });
+  });
 });
